requirements.txt
customtkinter
Pillow
pynput
pyperclip
packaging
